Rename: layout 폴더 생성 및 CommonPageLayout 컴포넌트 생성
