--- src/components/Layout.jsx
@@ -1,14 +1,0 @@
-import Footer from './Footer';
-import Header from './Header';
-
-function Layout({ children }) {
-  return (
-    <>
-      <Header></Header>
-      <main>{children}</main>
-      <Footer />
-    </>
-  );
-}
-
-export default Layout;
